Refactor page border utils
