+++ src/utils/nodes/page/attributes/pageBorders.ts
@@ -7,14 +7,14 @@
 import { EditorState, Transaction } from "@tiptap/pm/state";
 import { Dispatch, Editor } from "@tiptap/core";
 import { Node as PMNode } from "@tiptap/pm/model";
-import { PAGE_NODE_ATTR_KEYS } from "../constants/page";
-import { DEFAULT_PAGE_BORDER_CONFIG } from "../constants/pageBorders";
-import { BorderConfig } from "../types/page";
-import { Nullable } from "../types/record";
-import { px } from "./units";
-import { getPageAttributeByPageNum } from "./nodes/page/pageNumber";
-import { setPageNodePosSideConfig, updatePageSideConfig } from "./setSideConfig";
-import { MultiAxisSide } from "../types/page";
+import { PAGE_NODE_ATTR_KEYS } from "../../../../constants/page";
+import { DEFAULT_PAGE_BORDER_CONFIG } from "../../../../constants/pageBorders";
+import { BorderConfig } from "../../../../types/page";
+import { Nullable } from "../../../../types/record";
+import { px } from "../../../units";
+import { getPageAttributeByPageNum } from "../pageNumber";
+import { setPageNodePosSideConfig, updatePageSideConfig } from "../../../setSideConfig";
+import { MultiAxisSide } from "../../../../types/page";
 
 /**
  * Checks if a (single) border is valid.
